factor library add command to it’s own file
